fix: corred alb implementation
Associated JIRA issue is SOAR-2328.
Add unit tests to illustrate the the ALB processing isn't correct with respect to the CodeGenie codebase from which this was forked.
Update the implementation of alb.js so the tests pass.
Change the name of the package so InteliSys can use it independently from @codegenie